+++ handlers/redrabbit/redrabbit.go
@@ -8,9 +8,9 @@ import (
 	"strings"
 
 	"github.com/nyaruka/courier"
-	"github.com/nyaruka/courier/gsm7"
 	"github.com/nyaruka/courier/handlers"
 	"github.com/nyaruka/courier/utils"
+	"github.com/nyaruka/gocommon/gsm7"
 )
 
 var (